backend/requirements.txt
Flask
Flask-CORS
google-generativeai
python-docx
python-dotenv
